+++ tests/Integration/DSC_SqlRSSetup.Integration.Tests.ps1
@@ -56,6 +56,16 @@ BeforeAll {
             $script:mockSourceMediaUrl = 'https://download.microsoft.com/download/1/a/a/1aaa9177-3578-4931-b8f3-373b24f63342/SQLServerReportingServices.exe'
         }
 
+        if (Test-ContinuousIntegrationTaskCategory -Category 'Integration_SQL2022')
+        {
+            <#
+                The version below is what the MS download page said, but the .exe is
+                reporting 15.0.7842.32355 when used in the integration test.
+            #>
+            $script:mockSourceMediaDisplayName = 'Microsoft SQL Server 2022 Reporting Services (16.0.1113.11 - 11/23/2022)'
+            $script:mockSourceMediaUrl = 'https://download.microsoft.com/download/8/3/2/832616ff-af64-42b5-a0b1-5eb07f71dec9/SQLServerReportingServices.exe'
+        }
+
         Write-Verbose -Message ('Start downloading the {1} executable at {0}.' -f (Get-Date -Format 'yyyy-MM-dd hh:mm:ss'), $script:mockSourceMediaDisplayName) -Verbose
 
         Invoke-WebRequest -Uri $script:mockSourceMediaUrl -OutFile $ConfigurationData.AllNodes.SourcePath
@@ -92,7 +102,7 @@ AfterAll {
     Older versions of Reporting Services (eg. 2016) are integration tested in
     separate tests (part of resource SqlSetup).
 #>
-Describe "$($script:dscResourceName)_Integration" -Tag @('Integration_SQL2017', 'Integration_SQL2019') {
+Describe "$($script:dscResourceName)_Integration" -Tag @('Integration_SQL2017', 'Integration_SQL2019', 'Integration_SQL2022') {
     BeforeAll {
         $resourceId = "[$($script:dscResourceFriendlyName)]Integration_Test"
     }
@@ -156,6 +166,11 @@ Describe "$($script:dscResourceName)_Integration" -Tag @('Integration_SQL2017', 
             if (Test-ContinuousIntegrationTaskCategory -Category 'Integration_SQL2019')
             {
                 $resourceCurrentState.CurrentVersion | Should -BeGreaterThan ([System.Version] '15.0.0.0')
+            }
+
+            if (Test-ContinuousIntegrationTaskCategory -Category 'Integration_SQL2022')
+            {
+                $resourceCurrentState.CurrentVersion | Should -BeGreaterThan ([System.Version] '16.0.0.0')
             }
         }